+++ CHANGELOG.md
@@ -1,20 +1,30 @@
+# 2.1.0
+
+## New feature
+
+- Support ES6 import
+- Add typescript definition
+
 # 2.0.0
 
 ## Breaking change
 
 Instead of
+
 ```js
-const reverseDNSLookup = require('reverse-dns-lookup');
+const reverseDNSLookup = require("reverse-dns-lookup");
 const matches = await reverseDNSLookup(ip, domainA, domainB);
 ```
 
 Use
+
 ```js
-const { verify } = require('reverse-dns-lookup');
+const { verify } = require("reverse-dns-lookup");
 const matches = await verify(ip, domainA, domainB);
 ```
 
 "source" function is no longer available
 
 ## Internal change
+
 Use node's DNS module instead of unix commands.